pyproject.toml
[project]
name = "fika-prep"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.3.3",
    "ipykernel>=7.1.0",
    "notebook>=7.4.7",
    "supabase>=2.22.0",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
    "google-generativeai>=0.8.5",
    "osmnx>=2.0.7",
    "geopandas>=1.1.1",
    "psycopg2-binary>=2.9.11",
    "pyyaml>=6.0.3",
    "shapely>=2.1.2",
]

[tool.mypy]
ignore_missing_imports = true
